[Move `@kbn/config-schema` to server] `actions`
