+++ x-pack/plugins/actions/server/routes/connector/get_all/transforms/transform_connectors_response/v1.ts
@@ -6,7 +6,7 @@
  */
 
 import { ConnectorWithExtraFindData } from '../../../../../application/connector/types';
-import { AllConnectorsResponseV1 } from '../../../../../../common/routes/connector/response';
+import { AllConnectorsResponseV1 } from '../../../schemas/response';
 
 export const transformGetAllConnectorsResponse = (
   results: ConnectorWithExtraFindData[]
