+++ x-pack/plugins/actions/server/routes/connector/get_all_system/get_all_system.ts
@@ -6,7 +6,7 @@
  */
 
 import { IRouter } from '@kbn/core/server';
-import { AllConnectorsResponseV1 } from '../../../../common/routes/connector/response';
+import { AllConnectorsResponseV1 } from '../schemas/response';
 import { ActionsRequestHandlerContext } from '../../../types';
 import { INTERNAL_BASE_ACTION_API_PATH } from '../../../../common';
 import { ILicenseState } from '../../../lib';
