+++ x-pack/plugins/actions/server/routes/connector/list_types_system/list_types_system.ts
@@ -6,11 +6,11 @@
  */
 
 import { IRouter } from '@kbn/core/server';
-import { ConnectorTypesResponseV1 } from '../../../../common/routes/connector/response';
+import { ConnectorTypesResponseV1 } from '../schemas/response';
 import {
   connectorTypesQuerySchemaV1,
   ConnectorTypesRequestQueryV1,
-} from '../../../../common/routes/connector/apis/connector_types';
+} from '../schemas/apis/connector_types';
 import { ActionsRequestHandlerContext } from '../../../types';
 import { INTERNAL_BASE_ACTION_API_PATH } from '../../../../common';
 import { ILicenseState } from '../../../lib';
