+++ x-pack/plugins/actions/server/routes/connector/schemas/apis/get/schemas/v1.ts
